python/requirements.txt
certifi==2018.1.18
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.5.3
python-dateutil==2.6.1
six==1.11.0
smooch==2.9.0
urllib3==1.22
Werkzeug==0.14.1
